requirements.txt
streamlit
pandas
mysql-connector-python
matplotlib
